:lipstick: design: DrinksDetail 모바일 반응형 CSS 수정
