+++ client/src/components/Drinks/DrinksDetail/drinksdetailcontent/DrinksDetailSnacks.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import DrinksDetailSnack from "./DrinksDetailSnack";
-import DrinksDetailSnackTitle from "./DrinksDetailSnackTitle";
+import DrinksDetailSnack from "./drinksdetailsnacks/DrinksDetailSnack";
+import DrinksDetailSnackTitle from "./drinksdetailsnacks/DrinksDetailSnackTitle";
 import styled from "styled-components";
-import { IDrinksDetailProps } from '../../../util/interfaces/drinks.inerface'
+import { IDrinksDetailProps } from '../../../../util/interfaces/drinks.inerface'
 
 function DrinksDetailSnacks({ drinksDetail }: IDrinksDetailProps) {
