+++ client/src/components/Drinks/DrinksDetail/drinksdetailcontent/DrinksDetailTasting.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import DrinksDetailChart from "./DrinksDetailChart";
-import DrinksDetailChartTitle from "./DrinksDetailChartTitle";
+import DrinksDetailChart from "./drinksdetailtasting/DrinksDetailChart";
+import DrinksDetailChartTitle from "./drinksdetailtasting/DrinksDetailChartTitle";
 import styled from "styled-components";
-import { IDrinksDetailProps } from '../../../util/interfaces/drinks.inerface'
+import { IDrinksDetailProps } from '../../../../util/interfaces/drinks.inerface'
 
 function DrinksDetailTasting({ drinksDetail }: IDrinksDetailProps) {
   return (
