+++ client/src/components/Drinks/DrinksDetail/drinksdetailcontent/drinksdetailtasting/DrinksDetailChart.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import { IDrinksDetailProps } from '../../../util/interfaces/drinks.inerface'
+import { IDrinksDetailProps } from '../../../../../util/interfaces/drinks.inerface'
 
 import { Radar } from 'react-chartjs-2';
 import { ChartData, ChartOptions } from 'chart.js';
