+++ client/src/components/Drinks/DrinksDetail/drinksdetaillitem/DrinksDetailImg.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
-import Card from "../../UI/Card";
-import { IDrinksDetailProps } from '../../../util/interfaces/drinks.inerface'
+import Card from "../../../UI/Card";
+import { IDrinksDetailProps } from '../../../../util/interfaces/drinks.inerface'
 
 function DrinksDetailImg({ drinksDetail }: IDrinksDetailProps) {
   return (
