+++ client/src/components/Drinks/DrinksDetail/drinksdetaillitem/drinksdetailtxt/drinksdetailbox/DrinksDetailCount.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { HiOutlineChatBubbleOvalLeft } from "react-icons/hi2";
 import styled from 'styled-components';
-import { IDrinksDetailProps } from '../../../util/interfaces/drinks.inerface'
+import { IDrinksDetailProps } from '../../../../../../util/interfaces/drinks.inerface'
 
 
 function DrinksDetailCount({ drinksDetail }: IDrinksDetailProps) {
