+++ client/src/components/Drinks/DrinksDetail/drinksdetaillitem/drinksdetailtxt/drinksdetailbox/DrinksDetailTags.tsx
