+++ client/src/components/Drinks/DrinksList/DrinksContents.tsx
@@ -75,18 +75,18 @@ const ContentsContainer = styled.div`
   margin-bottom: var(--4x-large);
   justify-content: center;
 
-  @media only screen and (max-width: 768px) {
-    width: 100%;
-    height: auto;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
+    @media only screen and (max-width: 768px) {
+      width: 100%;
+      height: auto;
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+    }
 
-  @media only screen and (max-width: 450px) {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
+    @media only screen and (max-width: 450px) {
+      width: 100%;
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+    }
 `;
